Single preview chat + trace modal
