+++ packages/frontend/editor-ui/src/features/agents/__tests__/AgentSessionTimelinePanel.spec.ts
@@ -1,0 +1,94 @@
+/* eslint-disable import-x/no-extraneous-dependencies -- test-only patterns */
+import { describe, it, expect, vi, beforeEach } from 'vitest';
+import { mount, flushPromises } from '@vue/test-utils';
+import AgentSessionTimelinePanel from '../components/AgentSessionTimelinePanel.vue';
+import type { ThreadDetail } from '../composables/useAgentThreadsApi';
+
+const getThreadDetail = vi.fn();
+const showError = vi.fn();
+
+vi.mock('@/app/composables/useToast', () => ({
+	useToast: () => ({ showError }),
+}));
+
+vi.mock('@/features/agents/agentSessions.store', () => ({
+	useAgentSessionsStore: () => ({ getThreadDetail }),
+}));
+
+// Avoid the project-agents list dependency — the panel only reads `.value`.
+vi.mock('@/features/agents/composables/useSubAgentNames', () => ({
+	useSubAgentNames: () => ({ subAgentNameById: { value: new Map() } }),
+}));
+
+const stubs = {
+	SessionTimelineChart: { template: '<div data-test-id="chart-stub" />' },
+	SessionTimelineTable: { template: '<div data-test-id="table-stub" />' },
+	SessionEventFilter: { template: '<div data-test-id="filter-stub" />' },
+	SessionDetailPanel: { template: '<div data-test-id="detail-stub" />' },
+	N8nInput: { template: '<input data-test-id="search-stub" />' },
+	N8nIcon: { template: '<i />' },
+};
+
+const detail: ThreadDetail = {
+	thread: { id: 't1' } as ThreadDetail['thread'],
+	executions: [],
+};
+
+function mountPanel(props: Partial<{ projectId: string; agentId: string; threadId: string }> = {}) {
+	return mount(AgentSessionTimelinePanel, {
+		props: { projectId: 'p1', agentId: 'a1', threadId: 't1', ...props },
+		global: { stubs },
+	});
+}
+
+describe('AgentSessionTimelinePanel', () => {
+	beforeEach(() => {
+		vi.clearAllMocks();
+		getThreadDetail.mockResolvedValue(detail);
+	});
+
+	it('loads the thread detail on mount for its props', async () => {
+		mountPanel();
+		await flushPromises();
+
+		expect(getThreadDetail).toHaveBeenCalledWith('p1', 'a1', 't1');
+	});
+
+	it('emits loaded(null) at load start and the detail once resolved', async () => {
+		const wrapper = mountPanel();
+
+		// The immediate watch resets the header synchronously during mount.
+		expect(wrapper.emitted('loaded')?.[0]).toEqual([null]);
+
+		await flushPromises();
+
+		const events = wrapper.emitted('loaded') ?? [];
+		expect(events.at(-1)).toEqual([detail]);
+	});
+
+	it('renders the timeline table once loading finishes', async () => {
+		const wrapper = mountPanel();
+		await flushPromises();
+
+		expect(wrapper.find('[data-test-id="table-stub"]').exists()).toBe(true);
+	});
+
+	it('reloads when the threadId prop changes', async () => {
+		const wrapper = mountPanel();
+		await flushPromises();
+		expect(getThreadDetail).toHaveBeenCalledTimes(1);
+
+		await wrapper.setProps({ threadId: 't2' });
+		await flushPromises();
+
+		expect(getThreadDetail).toHaveBeenLastCalledWith('p1', 'a1', 't2');
+	});
+
+	it('surfaces a toast when the detail fails to load', async () => {
+		getThreadDetail.mockRejectedValueOnce(new Error('boom'));
+		mountPanel();
+		await flushPromises();
+
+		expect(showError).toHaveBeenCalled();
+	});
+});
