+++ packages/frontend/editor-ui/src/features/agents/components/AgentBuilderPreviewHeader.vue
@@ -11,17 +11,15 @@ import type { DropdownMenuItemProps } from '@n8n/design-system';
 import type { PathItem } from '@n8n/design-system/components/N8nBreadcrumbs/Breadcrumbs.vue';
 import { useI18n, type BaseTextKey } from '@n8n/i18n';
 import { computed } from 'vue';
-import { useRouter, type RouteLocationRaw } from 'vue-router';
 import KeyboardShortcutTooltip from '@/app/components/KeyboardShortcutTooltip.vue';
 import { useKeybindings } from '@/app/composables/useKeybindings';
-import { AGENT_SESSION_DETAIL_VIEW } from '@/features/agents/constants';
 
 const props = defineProps<{
 	breadcrumbItems: PathItem[];
 	sessionTitle: string;
-	sessionId?: string;
 	hasSession: boolean;
 	sessionOptions: Array<DropdownMenuItemProps<string>>;
+	traceOpen: boolean;
 }>();
 
 const emit = defineEmits<{
@@ -29,28 +27,24 @@ const emit = defineEmits<{
 	'session-select': [sessionId: string];
 	'new-chat': [];
 	'close-preview': [];
+	'toggle-trace': [];
 }>();
 
 const i18n = useI18n();
-const router = useRouter();
 
-const agentBreadcrumb = computed(() => props.breadcrumbItems[1]);
-const projectBreadcrumb = computed(() => props.breadcrumbItems[0]);
-const sessionRoute = computed<RouteLocationRaw | undefined>(() => {
-	if (!props.sessionId || !projectBreadcrumb.value || !agentBreadcrumb.value) return undefined;
-	return {
-		name: AGENT_SESSION_DETAIL_VIEW,
-		params: {
-			projectId: projectBreadcrumb.value.id,
-			agentId: agentBreadcrumb.value.id,
-			threadId: props.sessionId,
-		},
-	};
-});
+// The same button toggles the embedded session trace on/off, swapping its
+// label/icon between "view the trace" and "return to chat".
+const toggleTraceLabel = computed(() =>
+	i18n.baseText(
+		(props.traceOpen
+			? 'agents.builder.preview.backToChat'
+			: 'agents.builder.preview.viewSession') as BaseTextKey,
+	),
+);
+const toggleTraceIcon = computed(() => (props.traceOpen ? 'message-circle' : 'list-tree'));
 
-function openSession() {
-	if (!sessionRoute.value) return;
-	void router.push(sessionRoute.value);
+function toggleTrace() {
+	emit('toggle-trace');
 }
 
 function createNewChat() {
@@ -104,16 +98,16 @@ useKeybindings({
 		<div :class="$style.right">
 			<N8nTooltip v-if="props.hasSession" placement="bottom" :show-after="500">
 				<template #content>
-					{{ i18n.baseText('agents.builder.preview.viewSession' as BaseTextKey) }}
+					{{ toggleTraceLabel }}
 				</template>
 				<N8nButton
 					variant="ghost"
 					size="medium"
 					icon-size="large"
-					icon="list-tree"
-					:label="i18n.baseText('agents.builder.preview.viewSession' as BaseTextKey)"
+					:icon="toggleTraceIcon"
+					:label="toggleTraceLabel"
 					data-testid="agent-preview-view-session-btn"
-					@click="openSession"
+					@click="toggleTrace"
 				/>
 			</N8nTooltip>
 			<KeyboardShortcutTooltip
