+++ packages/frontend/editor-ui/src/features/agents/views/AgentBuilderView.vue
@@ -74,6 +74,7 @@ import AgentBuilderHeader from '../components/AgentBuilderHeader.vue';
 import AgentBuilderPreviewHeader from '../components/AgentBuilderPreviewHeader.vue';
 import AgentBuilderEditorColumn from '../components/AgentBuilderEditorColumn.vue';
 import AgentPreviewChatPage from '../components/AgentPreviewChatPage.vue';
+import AgentSessionTimelinePanel from '../components/AgentSessionTimelinePanel.vue';
 import AgentVersionHistoryPanel from '../components/VersionHistory/AgentVersionHistoryPanel.vue';
 import { useInstanceAiHandoff } from '@/features/ai/instanceAi/composables/useInstanceAiHandoff';
 import { useInstanceAiAvailable } from '@/features/ai/instanceAi/composables/useInstanceAiAvailability';
@@ -145,6 +146,11 @@ const { canUpdate: canEditAgent, canDelete: canDeleteAgent } = useAgentPermissio
 const effectiveCanEditAgent = computed(() => canEditAgent.value && !props.artifactEditingLocked);
 
 const isVersionHistoryOpen = ref(false);
+
+// Whether the preview shows the session trace (true) instead of the chat
+// (false). Local toggle only — no URL sync; the standalone session route
+// covers shareable deep-links.
+const isPreviewTraceOpen = ref(false);
 
 async function onSendPreviewToAssistant(executionId?: string) {
 	const threadId = effectiveSessionId.value;
@@ -1167,7 +1173,15 @@ watch(
 watch(isPreviewMode, (preview) => {
 	if (preview) {
 		bindPreviewSession();
+	} else {
+		isPreviewTraceOpen.value = false;
 	}
+});
+
+// Leaving a session (switching sessions) should drop back to the chat rather
+// than carry the trace-open state across to a different session.
+watch(effectiveSessionId, () => {
+	isPreviewTraceOpen.value = false;
 });
 
 function exitContinueMode() {
@@ -1291,13 +1305,14 @@ function onPreviewBreadcrumbSelect(item: PathItem) {
 			v-if="isPreviewMode"
 			:breadcrumb-items="previewBreadcrumbItems"
 			:session-title="currentSessionTitle"
-			:session-id="effectiveSessionId"
 			:has-session="currentSessionHasMessages"
 			:session-options="sessionOptions"
+			:trace-open="isPreviewTraceOpen"
 			@breadcrumb-select="onPreviewBreadcrumbSelect"
 			@session-select="onSessionPick"
 			@new-chat="onNewChat"
 			@close-preview="closePreview"
+			@toggle-trace="isPreviewTraceOpen = !isPreviewTraceOpen"
 		/>
 		<AgentBuilderHeader
 			v-else
@@ -1354,7 +1369,7 @@ function onPreviewBreadcrumbSelect(item: PathItem) {
 			</div>
 			<template v-else>
 				<AgentPreviewChatPage
-					v-if="isPreviewMode"
+					v-if="isPreviewMode && !isPreviewTraceOpen"
 					:initialized="initialized"
 					:project-id="projectId"
 					:agent-id="agentId"
@@ -1365,6 +1380,13 @@ function onPreviewBreadcrumbSelect(item: PathItem) {
 					:can-send-to-assistant="canSendPreviewToInstanceAi"
 					@continue-loaded="onContinueLoaded"
 					@send-to-assistant="onSendPreviewToAssistant"
+				/>
+
+				<AgentSessionTimelinePanel
+					v-else-if="isPreviewMode && isPreviewTraceOpen && effectiveSessionId"
+					:project-id="projectId"
+					:agent-id="agentId"
+					:thread-id="effectiveSessionId"
 				/>
 
 				<AgentBuilderEditorColumn
